DirectSearch.jl

DirectSearch.DefaultExtremeRef

DirectSearch.DefaultProgressiveRef

DirectSearch.AbstractMesh

DirectSearch.AbstractPoll

DirectSearch.AbstractProgressiveConstraint

DirectSearch.CollectionIndex

DirectSearch.ConstraintCollection

DirectSearch.ConstraintIndex

DirectSearch.ConstraintOutcome

DirectSearch.Constraints

DirectSearch.DSProblem

DirectSearch.ExtremeConstraint

DirectSearch.IterationOutcome

DirectSearch.LTMADS

DirectSearch.ProgressiveConstraint

DirectSearch.AddExtremeCollection

DirectSearch.AddExtremeConstraint

DirectSearch.AddExtremeConstraint

DirectSearch.AddProgressiveCollection

DirectSearch.AddProgressiveConstraint

DirectSearch.AddProgressiveConstraint

DirectSearch.CacheFilter

DirectSearch.CacheGet

DirectSearch.CacheGetRange

DirectSearch.CacheInitialPoint

DirectSearch.CachePush

DirectSearch.CachePush

DirectSearch.CacheQuery

DirectSearch.CacheRandomSample

DirectSearch.CollectionTypeCount

DirectSearch.ConstraintCollectionEvaluation

DirectSearch.ConstraintCollectionEvaluation

DirectSearch.ConstraintEvaluation

DirectSearch.ConstraintUpdate!

DirectSearch.EvaluatePoint!

DirectSearch.GenerateDirections

DirectSearch.GenerateSearchPoints

DirectSearch.GenerateSearchPoints

DirectSearch.GenerateSearchPoints

DirectSearch.GetHmaxSum

DirectSearch.MeshUpdate!

DirectSearch.Optimize!

DirectSearch.SetMaxEvals

DirectSearch.export_points

DirectSearch.function_evaluation

DirectSearch.function_evaluation

DirectSearch.h_max_update

DirectSearch.report

DirectSearch.report_finish

DirectSearch.DefaultExtremeRef — Constant

DefaultExtremeRef

The collection index that refers to the default location of extreme barrier constraints.

DirectSearch.DefaultProgressiveRef — Constant

DefaultProgressiveRef

The collection index that refers to the default location of progressive barrier constraints.

DirectSearch.DSProblem — Type

DSProblem{T}(N::Int; poll::AbstractPoll=LTMADS(N), search::AbstractSearch=NullSearch()) where T

Return a problem definition for an N dimensional problem. Specifiy the problem configuration with poll and search. By default LTMADS is used for the poll step, with nothing used for the search step.

DirectSearch.LTMADS — Type

LTMADS{T}()

Return an empty LTMADS object.

DirectSearch.AddExtremeCollection — Method

AddExtremeCollection(p::Constraints{T})::CollectionIndex where T

Instantiate a new constraint collection for extreme constraints. Returns an index that refers to the new collection.

DirectSearch.AddExtremeConstraint — Method

AddExtremeConstraint(p::AbstractProblem, c::Vector{Function})

Register a group of functions that define extreme barrier constraints. Return a vector of constraint indexes that refer to the constraints.

DirectSearch.AddExtremeConstraint — Method

AddExtremeConstraint(p::AbstractProblem, c::Function)::Tuple(ConstraintIndex, CollectionIndex)

Register a single function that defines an extreme barrier constraint. Return a constraint index.

The provided function should take a vector input and return a boolean value indicating if the constraint has been met or not.

The index argument can be specified to give a collection to add the constraint to. The specified collection must exist, and must be able to accept extreme barrier constraints.

DirectSearch.AddProgressiveCollection — Method

AddProgressiveCollection(p::Constraints{T}; h_max=Inf, h_max_update::Function=h_max_update, 
                         aggregator::Function=x->max(0,x)^2)::CollectionIndex where T

Instantiate a new constraint collection within the problem. Returns an index that refers to this new collection.

The default constraint settings match those from Audet & Dennis 2009:

h_max: Begins as infinity

h_max_update: Sets h_max to the largest valid h evaluation if an iteration is improving

aggregator: Creates h as ∑k(x) where k=max(0,x)^2

Note that the aggregator differs from that proposed by Audet & Dennis 2009 due to supporting multiple values of h_max at the same time.

DirectSearch.AddProgressiveConstraint — Method

AddProgressiveConstraint(p::AbstractProblem, c::Vector{Function})::Vector{Int}

Register a vector of functions that define a group of progressive barrier constraints. Return a vector of indices that refer to the constriants.

The provided functions should take a vector input and return a violation amount. The return type should be the same as the type that the problem is defined as (default is Float64).

DirectSearch.AddProgressiveConstraint — Method

AddProgressiveConstraint(p::AbstractProblem, c::Function)

Register a single function that defines a progressive barrier constraint. Return an index that refers to the constraint.

The provided function should take a vector input and return a violation amount. The return type should be the same as the type that the problem is defined as (default is Float64).

DirectSearch.Optimize! — Method

Optimize!(p::DSProblem)

Run the direct search algorithm on problem p.

p must have had its initial point set, which must be valid for whatever constraints have been given.

DirectSearch.report — Method

report(p::DSProblem; save::Union{Bool,String}=false)

Prints a summary of the problem to the console, or saves it as a text file.

save=true saves the report in the current REPL directory, but will error if the default file name report is taken. Otherwise give a path as an argument to use that name.

The report contains the solver configuration, the initial and final incumbent points, the initial and final cost, the duration, the number of iterations, and the current status.

DirectSearch.AbstractMesh — Type

abstract type AbstractMesh end

Parent type of any struct implementing the construction of a mesh. To maintain compatibility with other aspects of the package, the naming convention for variables within structs must be followed. These respect the notation used within Audet & Dennis 2006.

DirectSearch.AbstractPoll — Type

abstract type AbstractPollDirectionGenerator end

Parent type for any type used for implementing a direction generator.

Generally any direction generators should return vectors with unit length of one

DirectSearch.AbstractProgressiveConstraint — Type

AbstractProgressiveConstraint <: AbstractConstraint

Parent type for progressive contraints.

DirectSearch.CollectionIndex — Type

CollectionIndex

An Int wrapper that is used for indexing constraint collections within a Constraints object.

DirectSearch.ConstraintCollection — Type

ConstraintCollection{T,C}(hmax::T, hmax_update::Function, aggregator::Function) where {T,C<:AbstractConstraint}

Contains multiple constraints of the same type that have the same settings applied to them.

h_max is the initial hmax value. h_max_update is a function that should update h_max given an IterationOutcome value. aggregator is a function that will bring all constraint violations of a collection into a single h result.

Defaults for each of these values is set in the AddProgressiveCollection and AddExtremeCollection functions.

DirectSearch.ConstraintIndex — Type

CollectionIndex

An Int wrapper that is used for indexing constraints within a ConstraintCollection object.

DirectSearch.ConstraintOutcome — Type

@enum ConstraintOutcome

Has the values Feasible, WeakInfeasible, or StrongInfeasible to classify the outcome of the constraint evaluations of a point. A Feasible point meets the requirement of all constraints with no relaxation. A WeakInfeasible outcome requires one or more constraints to be relaxed, but less than their maximum amount. A StrongInfeasible outcome indicates one or more constraints was violated beyond its maximum relaxation amount.

DirectSearch.Constraints — Type

Constraints{T}() where T

Create an object that constains multiple constraint collection objects.

Upon creation Constraints is automativally populated with two constraint collections, an ExtremeCollection and a ProgressiveCollection.

DirectSearch.ExtremeConstraint — Type

ExtremeConstraint(f::Function)

Create an extreme barrier constraint. Function f should take a vector argument and return true or false to indicate if the vector meets the constraint.

DirectSearch.IterationOutcome — Type

@enum IterationOutcome

Has values Dominating, Improving, or Unsuccessful. Corresponding to the three iteration outcomes in progressive barrier direct search algorithms.

DirectSearch.ProgressiveConstraint — Type

ProgressiveConstraint(f::Function)

Create a progressive barrier constraint. Function f should take a vector argument and return a value in the range 0<=v<Inf to indicate the amount the constraint has been violated by the input vector. If input is a feasible point then the constraint function should return zero.

DirectSearch.CacheFilter — Method

CacheFilter(p::AbstractProblem{T}, points::Vector{T})::Tuple{Vector{Vector{T}},Vector{Vector{T}}} where T

Return a tuple where the first entry is the set of input points in the cache and the second is the set of input points not in the cache.

DirectSearch.CacheGet — Method

CacheGet(p::AbstractProblem, x::Vector)

Return the cost of point x in the cache of p. Does not check if x is in the cache, use CacheQuery to check.

DirectSearch.CacheGetRange — Method

CacheGetRange(p::AbstractProblem, points::Vector)::Vector{Vector}

Return a vector of costs corresponding to the vector of points.

DirectSearch.CacheInitialPoint — Method

CacheInitialPoint(p::AbstractProblem)

Return a tuple of the initial point added to the cache and its cost.

DirectSearch.CachePush — Method

CachePush(p::AbstractProblem, x::Vector, cost)

Add point x and its cost cost to the cache of p.

DirectSearch.CachePush — Method

CachePush(p::AbstractProblem)

Add the feasible and infeasible incumbent points (assuming neither are nothing) to the cache.

DirectSearch.CacheQuery — Method

CacheQuery(p::AbstractProblem, x::Vector)

Query the cache of p to find if it has a cost value for point x. Alias to haskey.

DirectSearch.CacheRandomSample — Method

CacheRandomSample(p::AbstractProblem, n::Int)

Returns a uniformly sampled collection of n points from the cache. Points can be repeated in the sample.

DirectSearch.CollectionTypeCount — Method

CollectionTypeCount(c::Constraints{T}, C::AbstractConstraint)::Int where T

Return the total number of constraints of type C that are stored in all collections.

DirectSearch.ConstraintCollectionEvaluation — Method

ConstraintCollectionEvaluation(collection::ConstraintCollection{T,ExtremeConstraint}, x::Vector{T})::ConstraintOutcome where T

Evalute every constraint within extreme constraint collection collection for point x.

If any constraint returns false then a StrongInfeasible result is returned. Otherwise a Feasible result is returned.

DirectSearch.ConstraintCollectionEvaluation — Method

ConstraintCollectionEvaluation(collection::ConstraintCollection{T,ProgressiveConstraint}, x::Vector{T})::ConstraintOutcome where T

Evalute every constraint within progressive constraint collection collection for point x.

If the aggregate value of the constraint evaluations exceeds h_max then a StrongInfeasible is returned. If the value is 0.0 then Feasible is returned. Otherwise WeakInfeasible is returned.

DirectSearch.ConstraintEvaluation — Method

ConstraintEvaluation(constraints::Constraints{T}, x::Vector{T})::Tuple{ConstraintOutcome,T} where T

Evaluate point x over all constraint collections in constraints. Returns a ConstraintOutcome indicating the result of the evaluation:

Feasible: x evaluated as feasible for all constraints (extreme and progressive barrier)

WeakInfeasible: x evaluated as feasible for all extreme barrier constraints, and had no progressive barrier constraint violations greater than h_max

StrongInfeasible: At least one extreme barrier constraint evaluated as infeasible, or at least one progressive barrier constraint had a violation greater than h_max

The second returned value is the sum of h_max values evaluated during the constraint checks.

DirectSearch.ConstraintUpdate! — Method

ConstraintUpdate!(c::Constraints, result::IterationOutcome)

Perform all necessary actions required to update a problem's constraints between iterations.

Currently will call h_max_update on each constraint.

DirectSearch.EvaluatePoint! — Method

EvaluatePoint!(p::DSProblem{T}, trial_points)::IterationOutcome where T

Determine whether the set of trial points result in a dominating, improving, or unsuccesful algorithm iteration. Update the feasible and infeasible incumbent points of p.

DirectSearch.GenerateDirections — Method

GenerateDirections(p::DSProblem, DG::LTMADS; maximal_basis=false)

Generates columns and forms a basis matrix for direction generation.

DirectSearch.GenerateSearchPoints — Method

GenerateSearchPoints(p::DSProblem)

Search method that returns an empty vector.

Use when no search method is desired.

DirectSearch.GenerateSearchPoints — Method

GenerateSearchPoints(p::DSProblem{T})::Vector{Vector{T}} where T

Alias for GenerateSearchPoints(p, NullSearch).

DirectSearch.GenerateSearchPoints — Method

GenerateSearchPoints(p::DSProblem{T}, ::RandomSearch)::Vector{Vector{T}} where T

Finds points that are Δᵐ distance from any point in the mesh in a uniformly random direction.

DirectSearch.GetHmaxSum — Method

GetHmaxSum(c::Constraints)

Return the sum of all h_max values for each collection in c.

DirectSearch.MeshUpdate! — Method

MeshUpdate!(mesh::Mesh, improvement_found::Bool)

Implements LTMADS update rule from Audet & Dennis 2006 pg. 203 adapted for progressive barrier constraints with Audet & Dennis 2009 expression 2.4

DirectSearch.SetMaxEvals — Method

SetMaxEvals(p::DSProblem, m::Int)

Set the maximum number of simultanious function evaluations that can be run. By default this will be set 1.

If the function evaluation is not overriden (e.g. for sending calculation to a cluster) then setting this to a number greater than your PC's number of threads will result in no improvement.

DirectSearch.export_points — Method

export_points(p::DSProblem)

Gives the trace of considered points during optimisation. Note that this is unavailable if tracking of points has been disabled

DirectSearch.function_evaluation — Method

function_evaluation(p::DSProblem{T}, trial_points::Vector{Vector{T}})::Vector{T} where T

Calculate the cost of the points in trial_points and return as a vector.

If the number of available workers is greater than one, and the maxsimultaniousevaluations value of p is greater than one then the calculation is distributed across several cores.

Currently, this has significant overheads and is much slower than evaluating in a single threaded manner on all testcases. This may give performance benefits when f is a heavy, single threaded operation.

If a specialised way of calling the function is needed then this function should be overriden, e.g.:

function DS.function_evaluation(p::DS.DSProblem{T}, trial_points::Vector{Vector{T}}) where T
	println("I am overriden")	
	return map(p.objective, trial_points)
end

DirectSearch.function_evaluation — Method

function_evaluation(p::DSProblem{T}, trial_point::Vector{T})::T where T

Evaluate a single trial point with the objective function of p.

By default calls the function with the trial point and returns the result. Override to provide custom evaluation behaviour.

DirectSearch.h_max_update — Method

h_max_update(c::ConstraintCollection, result::IterationOutcome)

Update h_max according to the LTMADS default in eqn 2.5, Audet & Dennis 2009

DirectSearch.report_finish — Method

report_finish(p::DSProblem)

If not silent, give a short printout of the result of the problem.

States the duration, final point, final cost, and reason for stopping. A more detailed report can be shown with report.