Public API

These functions and types implement the API of DirectSearch.jl.

Problem Configuration

DirectSearch.DSProblem — Type

DSProblem{T}(N::Int; poll::AbstractPoll=LTMADS{T}(),
                     search::AbstractSearch=NullSearch(),
                     objective::Union{Function,Nothing}=nothing,
                     initial_point::Vector=zeros(T, N),
                     iteration_limit::Int=1000,
                     )

Return a problem definition for an N dimensional problem.

poll and search specify the poll and search step algorithms to use. The default choices are (LTMADS)[@ref] and (NullSearch)[@ref] respectively.

Note that if working with Float64 (normally the case) then the type parameterisation can be ignored.

DirectSearch.SetObjective — Function

SetObjective(p::DSProblem, obj::Function)

Sets the target objective function to solve. obj should take a vector and return a single cost value.

DirectSearch.SetInitialPoint — Function

SetInitialPoint(p::DSProblem{T}, x::Vector{T}) where T

Set the initial incumbent point to x. This must be of the correct dimension. If using any extreme barrier constraints then it must also satisfy these constraints.

DirectSearch.Optimize! — Function

Optimize!(p::DSProblem)

Run the direct search algorithm on problem p.

p must have had its initial point and objective function set. If extreme barrier constraints have been set then the initial point must be value for those constraints.

DirectSearch.SetVariableRange — Function

SetVariableRange(p::DSProblem{T}, index::Int, l::T, u::T) where T

Set the expected range of the variable with index i to between lower (l) and upper (u) values. This does not create a constraint, and is only used for scaling when a variable varies with a significantly different scale to the others.

DirectSearch.SetVariableRanges — Function

SetVariableRanges(p::DSProblem{T}, l::Vector{T}, u::Vector{T}) where T

Call SetVariableRange for each variable. The vectors l and u should contain a lower and upper bound for each variable.

DirectSearch.SetGranularity — Function

SetGranularity(p::DSProblem{T}, index::Int, g::T) where T

Set the granularity of the variable with index i to 'g`.

DirectSearch.SetGranularities — Function

SetGranularities(p::DSProblem{T}, g::Vector{T}) where T

Call SetGranularity for each variable. The vector g should contain the granularity for each variable.

DirectSearch.SetSense — Function

SetSense(p::DSProblem, sense::ProblemSense )

Set the problem sense. Valid values for sense are DS.Min and DS.Max.

DirectSearch.SetMaxEvals — Function

SetMaxEvals(p::DSProblem, max::Bool=true)

Set to enable or disable parallel blackbox evaluations. The number of threads Julia was started with will be used.

Note that using parallel blackbox evaluations will only result in reduced runtime for problems that have long blackbox evaluations.

DirectSearch.SetOpportunisticEvaluation — Function

SetOpportunisticEvaluation(p::DSProblem; opportunistic::Bool=true)

Set/unset opportunistic evaluation (enables by default).

When using opportunistic evaluation the first allowable evaluated point with an improved cost is set as the new incumbent solution. If using progressive barrier constraints this point may be infeasible.

Search Stages

DirectSearch.NullSearch — Type

NullSearch()

Return no trial points for a search stage (ie, skips the search stage from running)

DirectSearch.RandomSearch — Type

RandomSearch(M::Int)

Return M randomly chosen trial points from the current mesh.

Poll Stages

DirectSearch.LTMADS — Type

LTMADS()

Return an empty LTMADS object.

LTMADS is a poll stage that creates a set of directions based on a semi-randomly generated lower triangular matrix. This randomness means that several runs of the algorithm may be needed to find a minimum.

DirectSearch.OrthoMADS — Type

OrthoMADS()

Return an empty OrthoMADS object.

OrthoMADS implements the poll set generation as described in Audet and Le Digabel 2015 Section 3.4.

Constraints

DirectSearch.AddExtremeConstraint — Method

AddExtremeConstraint(p::AbstractProblem, c::Function
                     index::CollectionIndex=CollectionIndex(1)
                    )::ConstraintIndex where T

Register a single function that defines an extreme barrier constraint. Return a constraint index.

The provided function should take a vector input and return a boolean or numeric value indicating if the constraint has been met or not. true or less than or equal to 0 indicates the constraint has been met. false or greater than 0 shows the constraint has been violated.

The index argument can be specified to give a collection to add the constraint to. The specified collection must exist, and must be able to accept extreme barrier constraints. If index is not specified then it is added to collection 1, the default extreme constraint collection.

DirectSearch.AddExtremeConstraint — Method

AddExtremeConstraint(p::AbstractProblem, c::Vector{Function}; index::CollectionIndex=CollectionIndex(1))

Register a group of functions that define extreme barrier constraints. Calls AddExtremeConstraint on each function individually.

DirectSearch.AddProgressiveConstraint — Method

AddProgressiveConstraint(p::AbstractProblem, c::Function; index::CollectionIndex=CollectionIndex(2))

Register a single function that defines a progressive barrier constraint. Return an index that refers to the constraint.

The provided function should take a vector input and return a numeric value indicating if the constraint has been met or not. Less than or equal to 0 indicates the constraint has been met. 0 shows the constraint has been violated.

The index argument can be specified to give a collection to add the constraint to. The specified collection must exist, and must be able to accept progressive barrier constraints. If index is not specified then it is added to collection 2, the default progressive barrier constraint collection.

DirectSearch.AddProgressiveConstraint — Method

AddProgressiveConstraint(p::AbstractProblem, c::Vector{Function})::Vector{Int}

Register a group of functions that define progressive barrier constraints. Calls AddProgressiveConstraint on each function individually.

DirectSearch.AddExtremeCollection — Method

AddExtremeCollection(p::Constraints{FT})::CollectionIndex where {FT<:AbstractFloat}

Instantiate a new constraint collection for extreme constraints. Returns an index that refers to the new collection.

DirectSearch.AddProgressiveCollection — Method

AddProgressiveCollection(p::Constraints{FT}; h_max=Inf, h_max_update::Function=h_max_update,
                         aggregator::Function=x->max(0,x)^2)::CollectionIndex where {FT<:AbstractFloat}

Instantiate a new constraint collection within the problem. Returns an index that refers to this new collection.

The default constraint settings match those from Audet & Dennis 2009:

h_max: Begins as infinity

h_max_update: Sets h_max to the largest valid h evaluation if an iteration is improving

aggregator: Creates h as $\sum k(x)$ where $k=\max(0,x)^2$

DirectSearch.DefaultExtremeRef — Constant

DefaultExtremeRef

The collection index that refers to the default location of extreme barrier constraints.

DirectSearch.DefaultProgressiveRef — Constant

DefaultProgressiveRef

The collection index that refers to the default location of progressive barrier constraints.

Cache

DirectSearch.CacheLoadJSON — Function

CacheLoadJSON(p::AbstractProblem{T}, path::String) where T

Load the costs from the provided JSON file to the cache. path can be a relative or absolute path and must contain the '.json' extension.

DirectSearch.CacheSaveJSON — Function

CacheSaveJSON(p::AbstractProblem{T}, filename::String) where T

Save the costs from the cache to JSON file with the name filename. Note that the filename should be without a file extension.

DirectSearch.CacheLoadJLD2 — Function

CacheLoadJLD2(p::AbstractProblem{T}, path::String, dataset::String="cache_costs") where T

Load the costs from the provided JLD2 file and dataset to the cache. path can be a relative or absolute path and must contain the '.jld2' extension. By default, it loads from the dataset named 'cache_costs'.

DirectSearch.CacheSaveJLD2 — Function

CacheSaveJLD2(p::AbstractProblem{T}, filename::String, dataset::String="cache_costs") where T

Save the costs from the cache to a JLD2 file with the name filename and to the dataset dataset. The default dataset is named 'cache_costs'. It is possible to write to multiple datasets in one file, however, datasets cannot be overwritten.

Stopping Conditions

Missing docstring.

Missing docstring for AddStoppingCondition. Check Documenter's build log for details.

DirectSearch.SetIterationLimit — Function

SetIterationLimit(p::DSProblem, i::Int)

Set the maximum number of iterations to i.

DirectSearch.BumpIterationLimit — Function

BumpIterationLimit(p::DSProblem, i::Int)

Increase the iteration limit by i.

DirectSearch.SetFunctionEvaluationLimit — Function

SetFunctionEvaluationLimit(p::DSProblem, i::Int)

Set the maximum number of function evaluations to i.

DirectSearch.BumpFunctionEvaluationLimit — Function

BumpFunctionEvaluationLimit(p::DSProblem, i::Int)

Increase the function evaluation limit by i.

DirectSearch.SetMinimumMeshSize — Function

SetMinimumMeshSize(p::DSProblem{T}, i::T) where T

Set the minimum mesh size for continuous variables.

DirectSearch.SetMinimumPollSize — Function

SetMinimumPollSize(p::DSProblem{T}, i::T) where T

Set the minimum poll size for continuous variables.

Missing docstring.

Missing docstring for RuntimeStoppingCondition. Check Documenter's build log for details.

Reporting

DirectSearch.ReportConfig — Function

ReportConfig(p::DSProblem)

Print the configuration options currently used by p.

DirectSearch.ReportStatus — Function

ReportStatus(p::DSProblem)

Print the current non-problem-specific status information of p.

DirectSearch.ReportProblem — Function

ReportProblem(p::DSProblem)

Print the current problem specific status information of p.

Base.print — Function

Base.print(p::DSProblem)

Print the output of ReportConfig, ReportStatus, and ReportProblem in a list.